remove apollo: remove LEAGUE_WEEK query. Remove queries.ts
