+++ apollo/resolvers.ts
@@ -5,241 +5,10 @@ import * as Sentry from "@sentry/nextjs";
 import { isUserAlreadyBelongToLeague } from "utils/isUserAlreadyBelongToLeague";
 import isUserAppliedToLeague from "utils/isUserAppliedToLeague";
 import { calculateCurrentGameweek } from "utils/calculateCurrentGameweek";
-import calculatePredictionScore from "utils/calculatePredictionScore";
 import isPastDeadline from "utils/isPastDeadline";
-import sortFixtures from "utils/sortFixtures";
-import Fixture from "src/types/Fixture";
-import League from "src/types/League";
-import User from "src/types/User";
-import calculateUsersLeaguePosition from "utils/calculateUsersLeaguePosition";
-import calculateWeeksUntilStart from "utils/calculateWeeksUntilStart";
-import calculateWeeksToGo from "utils/calculateWeeksToGo";
-import getWeeklyPoints from "utils/getWeeklyPoints";
 import dateScalar from "./scalars";
 
 const resolvers = {
-  Query: {
-    user: async (_, { userId }, context) => {
-      const id = userId ?? context?.user?.id;
-      if (!id) throw new ApolloError("User not found.");
-
-      const me = await prisma.user.findUnique({
-        where: { id },
-      });
-
-      return me;
-    },
-    currentGameweek: async (_, __) => {
-      // Cannot start in a past gameweek
-      const fixtures = await prisma.fixture.findMany({
-        select: {
-          id: true,
-          gameweek: true,
-          kickoff: true,
-        },
-      });
-
-      const currentGameweek = calculateCurrentGameweek(fixtures);
-
-      return currentGameweek;
-    },
-    fixturesWithPredictions: async (_, { leagueId, weekId }) => {
-      if (weekId < 1 || weekId > 38)
-        throw new UserInputError("Gameweek start week is not valid", {
-          argumentName: "gameweekStart",
-        });
-      /**
-       * TODO
-       * Change from:
-       * - Fetch users, their predictions and the fixtures those predictions belong to
-       * To:
-       * - Fetch fixtures, fetch users predictions, then match up.
-       */
-      const league = await prisma.league.findUnique({
-        where: {
-          id: leagueId,
-        },
-        include: {
-          users: {
-            select: {
-              id: true,
-              username: true,
-              predictions: {
-                select: {
-                  fixtureId: true,
-                  homeGoals: true,
-                  awayGoals: true,
-                  bigBoyBonus: true,
-                  fixture: {
-                    select: {
-                      gameweek: true,
-                      homeGoals: true,
-                      awayGoals: true,
-                    },
-                  },
-                },
-                where: {
-                  fixture: {
-                    gameweek: weekId,
-                  },
-                },
-              },
-            },
-          },
-        },
-      });
-      if (!league) throw new ApolloError("Cannot find league.");
-
-      const fixtures = await prisma.fixture.findMany({
-        where: {
-          gameweek: weekId,
-        },
-      });
-      if (!fixtures?.length) {
-        throw new ApolloError("No fixtures found for this gameweek.");
-      }
-
-      // Add missed predictions as 0-0
-      const usersWithAppendedPredictions = league.users.map((user) => {
-        const missingPredictions = [];
-        fixtures.forEach((fixture) => {
-          const isMissingPrediction =
-            user.predictions.findIndex((p) => p.fixtureId === fixture.id) ===
-            -1;
-
-          if (isMissingPrediction) {
-            missingPredictions.push({
-              fixtureId: fixture.id,
-              homeGoals: 0,
-              awayGoals: 0,
-              bigBoyBonus: false,
-              fixture: {
-                gameweek: weekId,
-                homeGoals: fixture.homeGoals,
-                awayGoals: fixture.awayGoals,
-              },
-            });
-          }
-        });
-
-        return {
-          ...user,
-          predictions: [...user.predictions, ...missingPredictions],
-        };
-      });
-
-      // Sort users
-      const users = usersWithAppendedPredictions.sort((a, b) => {
-        const totalPointsA = a.predictions.reduce((acc, cur) => {
-          const score = calculatePredictionScore(
-            [cur.homeGoals ?? 0, cur.awayGoals ?? 0, cur.bigBoyBonus ?? false],
-            [cur.fixture.homeGoals, cur.fixture.awayGoals]
-          );
-
-          return acc + score;
-        }, 0);
-
-        const totalPointsB = b.predictions.reduce((acc, cur) => {
-          const score = calculatePredictionScore(
-            [cur.homeGoals ?? 0, cur.awayGoals ?? 0, cur.bigBoyBonus ?? false],
-            [cur.fixture.homeGoals, cur.fixture.awayGoals]
-          );
-
-          return acc + score;
-        }, 0);
-
-        return totalPointsB - totalPointsA || (b.id > a.id ? 1 : -1);
-      });
-
-      const fixturesWithPredictions: Fixture[] = fixtures.map((fixture) => ({
-        ...fixture,
-        predictions: [],
-      }));
-
-      fixturesWithPredictions.forEach((fixture) => {
-        users.forEach((user) => {
-          const userPrediction = user.predictions.find(
-            (p) => p.fixtureId === fixture.id
-          );
-
-          if (userPrediction) {
-            const score = calculatePredictionScore(
-              [
-                userPrediction.homeGoals ?? 0,
-                userPrediction.awayGoals ?? 0,
-                userPrediction.bigBoyBonus ?? false,
-              ],
-              [
-                userPrediction.fixture.homeGoals,
-                userPrediction.fixture.awayGoals,
-              ]
-            );
-
-            fixture.predictions?.push({
-              user: {
-                id: user.id,
-              },
-              fixtureId: fixture.id,
-              homeGoals: userPrediction.homeGoals,
-              awayGoals: userPrediction.awayGoals,
-              bigBoyBonus: userPrediction.bigBoyBonus,
-              score,
-            });
-          } else {
-            fixture.predictions?.push({
-              user: {
-                id: user.id,
-              },
-              fixtureId: fixture.id,
-              homeGoals: null,
-              awayGoals: null,
-              bigBoyBonus: false,
-              score: null,
-            });
-          }
-        });
-      });
-
-      return {
-        fixtures: fixturesWithPredictions,
-      };
-    },
-    predictions: async (_, { weekId }, { user }) => {
-      const userId = user.id;
-
-      const predictions = await prisma.prediction.findMany({
-        where: {
-          AND: [
-            { userId },
-            {
-              fixture: {
-                gameweek: weekId,
-              },
-            },
-          ],
-        },
-        include: {
-          user: {
-            select: {
-              id: true,
-            },
-          },
-        },
-      });
-
-      return predictions;
-    },
-    league: async (_, { leagueId }) => {
-      const league = await prisma.league.findUnique({
-        where: {
-          id: leagueId,
-        },
-      });
-      if (!league) throw new ApolloError("Cannot find league.");
-
-      return league;
-    },
-  },
   Mutation: {
     updateUsername: async (_, { username }, { user: { id } }) => {
       if (username.length < 3 || username.length > 20)
@@ -535,217 +304,6 @@ const resolvers = {
           })
         ),
       };
-    },
-  },
-  User: {
-    username: async (parent: User) => {
-      const me = await prisma.user.findUnique({
-        where: { id: parent.id },
-      });
-
-      return me?.username || "";
-    },
-    leagues: async (user) => {
-      if (!user?.id) return [];
-
-      // Get all users leagues and all users that belong to those leagues, including their predictions
-      const userLeagues = await prisma.user.findUnique({
-        include: {
-          leagues: {
-            include: {
-              users: {
-                select: {
-                  id: true,
-                  predictions: {
-                    select: {
-                      fixtureId: true,
-                      homeGoals: true,
-                      awayGoals: true,
-                      bigBoyBonus: true,
-                      fixture: {
-                        select: {
-                          gameweek: true,
-                          homeGoals: true,
-                          awayGoals: true,
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-        where: {
-          id: user.id,
-        },
-      });
-      if (!userLeagues?.leagues.length) return [];
-
-      const fixtures = await prisma.fixture.findMany({
-        select: {
-          id: true,
-          gameweek: true,
-          kickoff: true,
-          homeGoals: true,
-          awayGoals: true,
-        },
-      });
-
-      const currentGameweek = calculateCurrentGameweek(fixtures);
-
-      const leagues = userLeagues.leagues
-        .map((league) => {
-          const users = league.users.map((u) => ({
-            ...u,
-            totalPoints: getWeeklyPoints(
-              fixtures,
-              u.predictions,
-              league.gameweekStart,
-              league.gameweekEnd
-            ).reduce((acc, cur) => acc + (cur.points || 0), 0),
-          }));
-
-          return {
-            leagueId: league.id,
-            leagueName: league.name,
-            gameweekStart: league.gameweekStart,
-            gameweekEnd: league.gameweekEnd,
-            weeksUntilStart: calculateWeeksUntilStart(
-              currentGameweek,
-              league.gameweekStart
-            ),
-            weeksToGo: calculateWeeksToGo(currentGameweek, league.gameweekEnd),
-            position: calculateUsersLeaguePosition(users, user.id),
-            numParticipants: league.users.length,
-            isActive: league.gameweekEnd >= currentGameweek,
-            users: league.users
-              .map((leagueUser) => ({
-                id: leagueUser.id,
-                totalPoints: leagueUser.predictions
-                  .filter(
-                    ({ fixture }) =>
-                      fixture.gameweek >= league.gameweekStart &&
-                      fixture.gameweek <= league.gameweekEnd
-                  )
-                  .reduce((totalPoints, prediction) => {
-                    const score = calculatePredictionScore(
-                      [
-                        prediction.homeGoals ?? 0,
-                        prediction.awayGoals ?? 0,
-                        prediction.bigBoyBonus ?? false,
-                      ],
-                      [
-                        prediction.fixture.homeGoals || 0,
-                        prediction.fixture.homeGoals || 0,
-                      ]
-                    );
-                    return totalPoints + score;
-                  }, 0),
-              }))
-              .sort((a, b) => b.totalPoints - a.totalPoints),
-          };
-        })
-        .sort((a, b) => a.gameweekStart - b.gameweekStart);
-
-      return leagues;
-    },
-  },
-  League: {
-    applicants: ({ applicants }: League) => {
-      const currentApplicants =
-        applicants?.filter((applicant) => applicant.status === "applied") || [];
-
-      return currentApplicants;
-    },
-    users: async ({ id }: League) => {
-      const league = await prisma.league.findUnique({
-        where: {
-          id,
-        },
-        include: {
-          users: {
-            select: {
-              id: true,
-              predictions: {
-                select: {
-                  fixtureId: true,
-                  homeGoals: true,
-                  awayGoals: true,
-                  bigBoyBonus: true,
-                  fixture: {
-                    select: {
-                      gameweek: true,
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      });
-      if (!league) throw new ApolloError("Cannot find league.");
-
-      const fixtures = await prisma.fixture.findMany({
-        select: {
-          id: true,
-          homeGoals: true,
-          awayGoals: true,
-          gameweek: true,
-          kickoff: true,
-        },
-      });
-
-      const users = league.users.map((user) => {
-        const predictions = user.predictions.filter(
-          (prediction) =>
-            prediction.fixture.gameweek >= league.gameweekStart &&
-            prediction.fixture.gameweek <= league.gameweekEnd
-        );
-
-        /**
-           The accumulator is an array of the form:
-           [{ week: 1, points: 0 }, { week: 2, points: 0 }, ... , { week: n, points: 0 }]
-
-           The callback function iterates through the fixtures, compares predictions against
-           each fixture to obtain the points for that fixture, then adds the points to the 
-           correct array index in the accumulator
-         */
-        const weeklyPoints = getWeeklyPoints(
-          fixtures,
-          predictions,
-          league.gameweekStart,
-          league.gameweekEnd
-        );
-
-        const totalPoints = weeklyPoints.reduce(
-          (acc, cur) => acc + (cur.points || 0),
-          0
-        );
-
-        return {
-          ...user,
-          weeklyPoints,
-          totalPoints,
-        };
-      });
-
-      const currentGameweek = calculateCurrentGameweek(fixtures);
-      const sortedUsers = users
-        .sort((a, b) => b.totalPoints - a.totalPoints || (b.id > a.id ? 1 : -1))
-        .map((user) => ({
-          ...user,
-          weeklyPoints: [
-            ...user.weeklyPoints.filter(
-              (weeklyPoints) =>
-                weeklyPoints.week >= league.gameweekStart &&
-                weeklyPoints.week <=
-                  Math.min(league.gameweekEnd, currentGameweek)
-            ),
-          ].reverse(),
-        }));
-
-      return sortedUsers;
     },
   },
   DateTime: dateScalar,
